requirements.txt
# ── Web Framework ──────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── Templating ─────────────────────────────────────────────────────
jinja2>=3.1.4
python-multipart>=0.0.9

# ── Database ───────────────────────────────────────────────────────
sqlalchemy>=2.0.30
pymysql>=1.1.0
cryptography>=42.0.7

# ── Redis ──────────────────────────────────────────────────────────
redis>=5.0.0                   # real-time counters + behaviour cache

# ── Rate Limiting ──────────────────────────────────────────────────
slowapi>=0.1.9                 # FastAPI-native rate limiting (uses limits)

# ── Security ───────────────────────────────────────────────────────
bcrypt>=4.1.3
python-jose[cryptography]>=3.3.0
secure>=0.3.0                  # HTTP secure headers middleware

# ── Configuration ──────────────────────────────────────────────────
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1

# ── Utilities ──────────────────────────────────────────────────────
httpx>=0.27.0
